+++ src/main/java/Application.java
@@ -1,11 +1,11 @@
-import domain.Team;
-import domain.board.Board;
-import domain.board.strategy.BoardCreateStrategy;
-import domain.board.strategy.MaSangMaSang;
-import domain.board.strategy.MaSangSangMa;
-import domain.board.strategy.SangMaMaSang;
-import domain.board.strategy.SangMaSangMa;
+import board.Board;
+import board.strategy.BoardCreateStrategy;
+import board.strategy.MaSangMaSang;
+import board.strategy.MaSangSangMa;
+import board.strategy.SangMaMaSang;
+import board.strategy.SangMaSangMa;
 import java.util.Map;
+import team.Team;
 import view.InputView;
 import view.InputView.CoordinatesPair;
 import view.OutputView;
