--- src/main/java/domain/piece/Cha.java
@@ -1,59 +1,0 @@
-package domain.piece;
-
-import domain.Coordinate;
-import domain.Team;
-import domain.board.Board;
-import java.util.HashSet;
-import java.util.Set;
-
-public class Cha extends Piece {
-
-    public Cha(Team team) {
-        super(team);
-    }
-
-    @Override
-    protected Set<Coordinate> findMovableCandidates(Coordinate departure) {
-        return departure.pickCrossCoordinates();
-    }
-
-    @Override
-    protected boolean canMoveConsideringObstacles(Board board, Coordinate departure, Coordinate arrival) {
-        return findPaths(departure, arrival)
-                .stream()
-                .noneMatch(board::isExistence);
-    }
-
-    @Override
-    protected Set<Coordinate> findPaths(Coordinate departure, Coordinate arrival) {
-        int dx = arrival.getX() - departure.getX();
-        int dy = arrival.getY() - departure.getY();
-
-        Set<Coordinate> coordinates = new HashSet<>();
-        if (dx == 0 && dy > 0) { // 아래
-            for (int y = departure.getY() + 1; y < arrival.getY(); y++) {
-                coordinates.add(new Coordinate(departure.getX(), y));
-            }
-            return coordinates;
-        }
-        if (dx == 0 && dy < 0) { // 위
-            for (int y = departure.getY() - 1; y > arrival.getY(); y--) {
-                coordinates.add(new Coordinate(departure.getX(), y));
-            }
-            return coordinates;
-        }
-        if (dx > 0 && dy == 0) { // 오른쪽
-            for (int x = departure.getX() + 1; x < arrival.getX(); x++) {
-                coordinates.add(new Coordinate(x, departure.getY()));
-            }
-            return coordinates;
-        }
-        if (dx < 0 && dy == 0) { // 왼쪽
-            for (int x = departure.getX() - 1; x > arrival.getX(); x--) {
-                coordinates.add(new Coordinate(x, departure.getY()));
-            }
-            return coordinates;
-        }
-        throw new IllegalStateException("유효하지 않은 좌표입니다.");
-    }
-}
