+++ src/main/java/team/Team.java
@@ -1,0 +1,18 @@
+package team;
+
+public enum Team {
+
+    HAN(1),
+    CHO(-1),
+    ;
+
+    private final int direction;
+
+    Team(int direction) {
+        this.direction = direction;
+    }
+
+    public int getDirection() {
+        return direction;
+    }
+}
